Stuffed Lexical ANalyzer in class
